Make n-iid's sample much, much faster
The previous version used build-vector with a lambda to sample every
element of the vector, which was horrifically slow -- I suspect because
of Typed Racket overhead and the overhead of calling a lambda repeatedly
instead of just sampling the right number of values in the first
place. This version is about ten times faster for the
two-sample-power.rkt example.